Update Exploratory Programming 2.ipynb
This is my code that checks to see if the data is a list or not. If it is, it runs a check on the list to count up the amount of positive numbers in the list and print it to the console. Two inputs are shown, one as a list, and one as a string to illustrate that the TypeError is raised.